+++ wowusky/__init__.py
@@ -1,4 +1,4 @@
 """wowusky — minimalist World of Warcraft addon manager for Linux."""
 
-__version__ = "0.5.2"
+__version__ = "0.5.3"
 APP_NAME    = "wowusky"